moved samples/dev/tests to examples folder
