+++ examples/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './src/App';
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(<App/>, document.getElementById('react-light-components'));
