+++ examples/package.json
